+++ CoinWeight/gameview.hpp
@@ -19,9 +19,9 @@ public:
     virtual void drawMainScreen(int screenHighlight) = 0;
     virtual void drawInstructionScreen() = 0;
     virtual void drawCreditScreen() = 0;
-    virtual void drawGameOptionScreen(int screenHighlight, int numOfCoins,
+    virtual void drawGameOptionScreen(int screenHighlight, size_t numOfCoins,
                     std::string gameLevel, bool isHuman) = 0;
-    virtual void drawGamePlayScreen(int numOfCoins, std::vector<int> coinStates,
+    virtual void drawGamePlayScreen(size_t numOfCoins, std::vector<int> coinStates,
                     std::vector<std::pair<Weighing, int>> gameHistory) = 0;
     virtual void drawGameOverScreen() = 0;
 };
